Prefer responses endpoint for GPT models
